Moved the MultiProcessRedisLock class to its own file so that users choose to use redis.
