+++ tests/tests/test_locks.py
@@ -1,11 +1,11 @@
 from django.test import override_settings
 import pytest
 
-from idempotency_key import locks
+from idempotency_key.locks import basic, redis
 
 
 def test_single_thread_lock():
-    obj = locks.ThreadLock()
+    obj = basic.ThreadLock()
     assert obj.acquire() is True
     assert obj.acquire() is False
     obj.release()
@@ -15,7 +15,7 @@ def test_single_thread_lock():
 
 @override_settings(IDEMPOTENCY_KEY={"LOCK": {"LOCATION": "Redis://localhost"}})
 def test_multi_process_lock_only_host():
-    obj = locks.MultiProcessRedisLock()
+    obj = redis.MultiProcessRedisLock()
     assert obj.acquire() is True
     assert obj.acquire() is False
     obj.release()
@@ -25,7 +25,7 @@ def test_multi_process_lock_only_host():
 
 @override_settings(IDEMPOTENCY_KEY={"LOCK": {"LOCATION": "Redis://localhost:6379/1"}})
 def test_multi_process_lock_host_and_port():
-    obj = locks.MultiProcessRedisLock()
+    obj = redis.MultiProcessRedisLock()
     assert obj.acquire() is True
     assert obj.acquire() is False
     obj.release()
@@ -36,10 +36,10 @@ def test_multi_process_lock_host_and_port():
 @override_settings(IDEMPOTENCY_KEY={"LOCK": {"LOCATION": ""}})
 def test_multi_process_lock_empty_string_must_be_set():
     with pytest.raises(ValueError):
-        locks.MultiProcessRedisLock()
+        redis.MultiProcessRedisLock()
 
 
 @override_settings(IDEMPOTENCY_KEY={"LOCK": {"LOCATION": None}})
 def test_multi_process_lock_null_must_be_set():
     with pytest.raises(ValueError):
-        locks.MultiProcessRedisLock()
+        redis.MultiProcessRedisLock()
